requirements.txt
fastapi
uvicorn
streamlit
chromadb
sentence-transformers
openai
pyjwt
requests
pydantic
